requirements.txt
# Core requirements for document processor
pyyaml>=6.0.1
jinja2>=3.1.2
markdown>=3.4.1
requests>=2.28.0
msgraph-core>=0.2.2

# Microsoft Authentication Library
msal>=1.20.0

# Azure Key Vault dependencies
azure-identity>=1.12.0
azure-keyvault-secrets>=4.6.0

# Optional: For loading environment variables
python-dotenv>=1.0.0
